refactor(mobile): inline EnrichedBook type, delete library-filters.ts
- library-filters.ts held only the EnrichedBook interface after ~250 LOC
  of unused v2 client-filter helpers were removed earlier
- Inline the interface into its sole consumer use-all-books.ts
- Delete now-empty library-filters.ts
